Fixed Query Error
Sometimes an uncaught exception would occur when an input is provided to the Nominatim web service that for some reason spits back an html page instead of a json file. This patch adds a new test to GeoLocatorTest and fixes the issue by pattern matching for the DOCTYPE declaration of HTMLs and escaping if that is found.