Fixed overlap issue and new picture
Fixed the issue where the header content would overlap with the navigation bar when viewed from a phone. The css now checks the screen size and if a phone size screen is used then the content is made to take a limited portion of the screen and a scroll bar is added.  Also a new background image was added along with a clickable footer for citation.